+++ R/zzz.R
@@ -13,5 +13,6 @@
         pop[["whoami"]] <- adc
         options(packager = pop)
     }
+    fritools::run_r_tests_for_known_hosts()
     return(invisible(NULL))
 }