+++ inst/templates/man-roxygen/return_boolean.R
@@ -1,0 +1,4 @@
+#' @return \code{\link[base:TRUE]{TRUE}} on success, 
+#' \code{\link[base:FALSE]{FALSE}} otherwise.
+
+