+++ inst/templates/zzz.R
@@ -1,0 +1,4 @@
+.onLoad <- function(libname, pkgname) {
+    fritools::run_r_tests_for_known_hosts()
+    return(invisible(NULL))
+}